requirements.txt
tensorflow>=2.12.0
opencv-python>=4.8.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
scikit-learn>=1.0.0
streamlit>=1.25.0
streamlit-webrtc>=0.45.0
